+++ sl-map-apis/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.2
+
+add methods to the `MapTileCache` to check if a map tile or a region exist
+add the option to fill missing map tiles in `Map` with a color
+add the option to fill missing regions in lower detail map tiles with a color
+
 ## 0.1.1
 
 update sl-types dependency to 0.1.1