+++ sl-map-cli/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.2
+
+allow optionally filling in missing regions with a color
+allow optionally filling in missing map tiles with a color
+allow saving of the USB notecard map without a route to an extra file
+before the route is drawn on it
+
 ## 0.1.1
 
 Print PPS HUD config string and size recommendations when generating a map